:recycle: Extract job_step.go from progress_notification.go
